Move some code from merge.py to profilers/perf.py file
