+++ gprofiler/merge.py
@@ -5,12 +5,11 @@
 import json
 import math
 import random
-import re
 from collections import Counter, defaultdict
 from dataclasses import dataclass
 from datetime import datetime
 from pathlib import Path
-from typing import Any, Dict, Iterable, List, Optional, Tuple
+from typing import Any, Dict, List, Optional, Tuple
 
 from granulate_utils.metadata import Metadata
 
@@ -29,16 +28,9 @@ from gprofiler.utils import merge_dicts
 
 logger = get_logger_adapter(__name__)
 
-SAMPLE_REGEX = re.compile(
-    r"\s*(?P<comm>.+?)\s+(?P<pid>[\d-]+)/(?P<tid>[\d-]+)(?:\s+\[(?P<cpu>\d+)])?\s+(?P<time>\d+\.\d+):\s+"
-    r"(?:(?P<freq>\d+)\s+)?(?P<event_family>[\w\-_/]+):(?:(?P<event>[\w-]+):)?(?P<suffix>[^\n]*)(?:\n(?P<stack>.*))?",
-    re.MULTILINE | re.DOTALL,
-)
-
 # ffffffff81082227 mmput+0x57 ([kernel.kallsyms])
 # 0 [unknown] ([unknown])
 # 7fe48f00faff __poll+0x4f (/lib/x86_64-linux-gnu/libc-2.31.so)
-FRAME_REGEX = re.compile(r"^\s*[0-9a-f]+ (.*?) \(\[?(.*?)\]?\)$")
 
 
 def parse_one_collapsed(collapsed: str, add_comm: Optional[str] = None) -> StackToSampleCount:
@@ -97,78 +89,6 @@ def parse_many_collapsed(text: str) -> ProcessToStackSampleCounters:
     return results
 
 
-def _collapse_stack(comm: str, stack: str, insert_dso_name: bool = False) -> str:
-    """
-    Collapse a single stack from "perf".
-    """
-    funcs = [comm]
-    for line in reversed(stack.splitlines()):
-        m = FRAME_REGEX.match(line)
-        assert m is not None, f"bad line: {line}"
-        sym, dso = m.groups()
-        sym = sym.split("+")[0]  # strip the offset part.
-        if sym == "[unknown]" and dso != "unknown":
-            sym = f"({dso})"
-        # append kernel annotation
-        elif "kernel" in dso or "vmlinux" in dso:
-            sym += "_[k]"
-        elif insert_dso_name:
-            sym += f" ({dso})"
-        funcs.append(sym)
-    return ";".join(funcs)
-
-
-def merge_global_perfs(
-    raw_fp_perf: Optional[str], raw_dwarf_perf: Optional[str], insert_dso_name: bool = False
-) -> ProcessToStackSampleCounters:
-    fp_perf = _parse_perf_script(raw_fp_perf, insert_dso_name)
-    dwarf_perf = _parse_perf_script(raw_dwarf_perf, insert_dso_name)
-
-    if raw_fp_perf is None:
-        return dwarf_perf
-    elif raw_dwarf_perf is None:
-        return fp_perf
-
-    total_fp_samples = sum([sum(stacks.values()) for stacks in fp_perf.values()])
-    total_dwarf_samples = sum([sum(stacks.values()) for stacks in dwarf_perf.values()])
-    fp_to_dwarf_sample_ratio = total_fp_samples / total_dwarf_samples
-
-    # The FP perf is used here as the "main" perf, to which the DWARF perf is scaled.
-    merged_pid_to_stacks_counters: ProcessToStackSampleCounters = defaultdict(Counter)
-    add_highest_avg_depth_stacks_per_process(
-        dwarf_perf, fp_perf, fp_to_dwarf_sample_ratio, merged_pid_to_stacks_counters
-    )
-    total_merged_samples = sum([sum(stacks.values()) for stacks in merged_pid_to_stacks_counters.values()])
-    logger.debug(
-        f"Total FP samples: {total_fp_samples}; Total DWARF samples: {total_dwarf_samples}; "
-        f"FP to DWARF ratio: {fp_to_dwarf_sample_ratio}; Total merged samples: {total_merged_samples}"
-    )
-    return merged_pid_to_stacks_counters
-
-
-def add_highest_avg_depth_stacks_per_process(
-    dwarf_perf: ProcessToStackSampleCounters,
-    fp_perf: ProcessToStackSampleCounters,
-    fp_to_dwarf_sample_ratio: float,
-    merged_pid_to_stacks_counters: ProcessToStackSampleCounters,
-) -> None:
-    for pid, fp_collapsed_stacks_counters in fp_perf.items():
-        if pid not in dwarf_perf:
-            merged_pid_to_stacks_counters[pid] = fp_collapsed_stacks_counters
-            continue
-
-        fp_frame_count_average = get_average_frame_count(fp_collapsed_stacks_counters.keys())
-        dwarf_collapsed_stacks_counters = dwarf_perf[pid]
-        dwarf_frame_count_average = get_average_frame_count(dwarf_collapsed_stacks_counters.keys())
-        if fp_frame_count_average > dwarf_frame_count_average:
-            merged_pid_to_stacks_counters[pid] = fp_collapsed_stacks_counters
-        else:
-            dwarf_collapsed_stacks_counters = scale_sample_counts(
-                dwarf_collapsed_stacks_counters, fp_to_dwarf_sample_ratio
-            )
-            merged_pid_to_stacks_counters[pid] = dwarf_collapsed_stacks_counters
-
-
 def scale_sample_counts(stacks: StackToSampleCount, ratio: float) -> StackToSampleCount:
     if ratio == 1:
         return stacks
@@ -184,57 +104,6 @@ def scale_sample_counts(stacks: StackToSampleCount, ratio: float) -> StackToSamp
         if scaled_value != 0:
             scaled_stacks[stack] = scaled_value
     return scaled_stacks
-
-
-def get_average_frame_count(samples: Iterable[str]) -> float:
-    """
-    Get the average frame count for all samples.
-    Avoids counting kernel frames because this function is used to determine whether FP stacks
-    or DWARF stacks are to be used. FP stacks are collected regardless of FP or DWARF, so we don't
-    count them in this heuristic.
-    """
-    frame_count_per_samples = []
-    for sample in samples:
-        kernel_split = sample.split("_[k];", 1)
-        if len(kernel_split) == 1:
-            kernel_split = sample.split("_[k] ", 1)
-
-        # Do we have any kernel frames in this sample?
-        if len(kernel_split) > 1:
-            # example: "a;b;c;d_[k];e_[k] 1" should return the same value as "a;b;c 1", so we don't
-            # add 1 to the frames count like we do in the other branch.
-            frame_count_per_samples.append(kernel_split[0].count(";"))
-        else:
-            # no kernel frames, so e.g "a;b;c 1" and frame count is one more than ";" count.
-            frame_count_per_samples.append(kernel_split[0].count(";") + 1)
-    return sum(frame_count_per_samples) / len(frame_count_per_samples)
-
-
-def _parse_perf_script(script: Optional[str], insert_dso_name: bool = False) -> ProcessToStackSampleCounters:
-    pid_to_collapsed_stacks_counters: ProcessToStackSampleCounters = defaultdict(Counter)
-
-    if script is None:
-        return pid_to_collapsed_stacks_counters
-
-    for sample in script.split("\n\n"):
-        try:
-            if sample.strip() == "":
-                continue
-            if sample.startswith("#"):
-                continue
-            match = SAMPLE_REGEX.match(sample)
-            if match is None:
-                raise Exception("Failed to match sample")
-            sample_dict = match.groupdict()
-
-            pid = int(sample_dict["pid"])
-            comm = sample_dict["comm"]
-            stack = sample_dict["stack"]
-            if stack is not None:
-                pid_to_collapsed_stacks_counters[pid][_collapse_stack(comm, stack, insert_dso_name)] += 1
-        except Exception:
-            logger.exception(f"Error processing sample: {sample}")
-    return pid_to_collapsed_stacks_counters
 
 
 def _make_profile_metadata(
